move mock routing tests to proper directory
