+++ routing/mock/routing.go
@@ -1,4 +1,4 @@
-package mockrouter
+package mock
 
 import (
 	"errors"
